requirements.txt
Django==5.0.4
django-cleanup==8.1.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.3.0